requirements.txt
PyYAML==5.1.2
pytest==5.1.2
virtualenv==16.7.5
waitress==1.3.1
python-snappy==0.5.4
Flask==1.1.1
Flask-RESTful==0.3.7
numpy==1.16.4
scipy==1.3.1
seaborn==0.9.0
netCDF4==1.5.3
